Provide documentation for new @EncodeDefault and @JsonClassDiscriminator annotations
Fixes #1657
