+++ guide/example/example-basic-03.kt
@@ -4,7 +4,7 @@ package example.exampleBasic03
 import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
-@Serializable 
+@Serializable
 data class Project(val name: String, val language: String)
 
 fun main() {
