+++ guide/example/example-classes-04.kt
@@ -4,7 +4,7 @@ package example.exampleClasses04
 import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
-@Serializable 
+@Serializable
 data class Project(val name: String, val language: String)
 
 fun main() {
