+++ guide/example/example-classes-08.kt
@@ -4,7 +4,7 @@ package example.exampleClasses08
 import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
-@Serializable 
+@Serializable
 data class Project(val name: String, @Transient val language: String = "Kotlin")
 
 fun main() {
