+++ guide/example/example-classes-10.kt
@@ -5,9 +5,21 @@ import kotlinx.serialization.*
 import kotlinx.serialization.json.*
 
 @Serializable
-class Project(val name: String, val renamedTo: String? = null)
+data class Project(
+    val name: String,
+    @EncodeDefault val language: String = "Kotlin"
+)
+
+
+@Serializable
+data class User(
+    val name: String,
+    @EncodeDefault(EncodeDefault.Mode.NEVER) val projects: List<Project> = emptyList()
+)
 
 fun main() {
-    val data = Project("kotlinx.serialization")
-    println(Json.encodeToString(data))
+    val userA = User("Alice", listOf(Project("kotlinx.serialization")))
+    val userB = User("Bob")
+    println(Json.encodeToString(userA))
+    println(Json.encodeToString(userB))
 }
